+++ src/components/02-organisms/Lomake/index.js
@@ -175,7 +175,6 @@ const Lomake = React.memo(
     ]);
 
     if (Array.isArray(lomake) && lomake.length) {
-      console.info(anchor, changeObjects);
       return (
         <div>
           {code || formTitle ? (
