+++ src/scenes/Koulutusmuodot/AmmatillinenKoulutus/EsittelijatMuutospyynto.js
@@ -165,7 +165,7 @@ const EsittelijatMuutospyynto = ({
           formTitle={sectionHeadings.opetusJaTutkintokieli.title}
           isPreviewModeOn={false}
           isRowExpanded={true}
-          mode="modification"
+          mode={mode}
           path={constants.formLocation.opetuskielet}
           rowTitle={intl.formatMessage(wizardMessages.teachingLanguages)}
           showCategoryTitles={true}
