+++ src/services/lomakkeet/tutkinnot/index.js
@@ -153,7 +153,6 @@ async function getModificationForm(
               tutkintoCategory,
               osaamisalaCategories
             );
-            console.info(isAddition, reasoningLomakeosio);
           }
 
           return {
